Added a PHP helper script that grabs Shaarli sites from various sources
Currently thoses sites are read from the Shaarli RSS site (URI is in PHP script).
Notice that, to avoid rewriting sites, I've taken the freedom to add to each theme
 a config.ini file that will give the script the source url for the shaarli site.
This way, two runs of that script will only download styles that are not yet defined
in themes.

To avoid automatic procedure, also notice all styles are downloaded to a "new_themes" directory.
It is the script runner duty to copy by hand the scripts to a good location.